+++ app/models/validation/validation.rb
@@ -15,8 +15,8 @@ module Validation
     'signature' => Validation::Validators::Signature,
     # 'string' => Validation::Validators::String,
     # 'text' => Validation::Validators::Text,
-    'time' => Validation::Validators::TimeOfDay
-    # 'time_duration' => Validation::Validators::TimeDuration
+    'time' => Validation::Validators::TimeOfDay,
+    'time_duration' => Validation::Validators::TimeDuration
   }
 
   def self.for(object)