+++ lib/date_and_time_parser.rb
@@ -65,6 +65,41 @@ module DateAndTimeParser
     default_time
   end
 
+  def parse_time_duration(time_duration_string)
+    hash = nil
+    sections = time_duration_string.to_s.split(':')
+    hours = parse_integer(sections[0])
+    minutes = parse_integer(sections[1])
+    seconds = parse_integer(sections[2])
+    if hours || minutes || seconds
+      hash = {}
+      hash[:hours]   = hours   || 0
+      hash[:minutes] = minutes || 0
+      hash[:seconds] = seconds || 0
+    end
+    hash
+  end
+
+  def parse_time_duration_from_hash(time_duration_hash)
+    if time_duration_hash.is_a? Hash
+      Rails.logger.debug time_duration_hash.inspect
+      hours = parse_integer(time_duration_hash[:hours])
+      minutes = parse_integer(time_duration_hash[:minutes])
+      seconds = parse_integer(time_duration_hash[:seconds])
+      Rails.logger.debug "#{hours}:#{minutes}:#{seconds}"
+      parse_time_duration("#{hours}:#{minutes}:#{seconds}")
+    else
+      parse_time_duration('')
+    end
+  end
+
+  def parse_time_duration_from_hash_to_s(time_duration_hash, default_time_duration: '')
+    hash = parse_time_duration_from_hash(time_duration_hash)
+    "#{hash[:hours]}:#{hash[:minutes]}:#{hash[:seconds]}"
+  rescue
+    default_time_duration
+  end
+
   def parse_integer(string)
     Integer(format('%g', string))
   rescue